Phase3/Code/requirements.txt
pandas
scipy
numpy
scikit-image==0.15.0
sklearn
pillow
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
